macOS: revert c99f330: user_config_dir should be the same as user_data_dir...
and site_config_dir should be the same as site_data_dir, unless of
course the relevant XDG_* env vars are set and non-empty, in which case
we follow XDG, which is what this fork sets out to do in the first
place.

c99f330 changed user_config_dir to ~/Library/Preferences/<AppName> and
site_config_dir to /Library/Preferences/<AppName>, which is complete
garbage. The Preferences directory are specifically for plists. If
you're reading/writing a plist and you need appdirs to tell you where to
write, you're an idiot; if you're not reading/writing a plist, but an
ini/json/yaml/whatever config file, and you're looking into Preferences,
you're also an idiot.